Fix tests
Not setting APP_ENV in tests causes it to default to "production", and
when that is the artisan migrate command will require confirmation
before nuking the db. In CI it aborts (likely because stdin isn't a
TTY), while locally it would just hang (waiting for input, but not
having shown any output).