Fix Dredd invocation
In the current version Dredd requires significant CLI configuration
to run. This is probably due to a previous dredd.yml file which was
not committed.

Here we add a new version of the file with configuration options that
should work.

The current implementation also seems to suffer from a problem where
the built-in PHP server started by Dredd in dredd-server.sh could not
be shut down automatically on OSX.

To try to address this we start the PHP server directly from Dredd
and configure the application and seed the database using a 
before-all hook.

To avoid having to clean up the database we switch to an in-memory
version of SQLite. Note that this may make debugging a bit more
difficult if a developer wants to accesses the contents of the 
database after a test.