Setup middleware to validate tokens against Adgangsplatformen
The is the authentication method that will be used for non-test
environments.

The OAuth library for Adgangsplatformen provides PSR-15 middleware
for validating access tokens and mapping these to resource owners.

The Laravel PSR-15 bridge provides the simplest way to use PSR-15
middleware in Lumen as it simply wraps an PSR-15 middleware object.
The more popular alternative https://github.com/jshannon63/laravel-psr15-middleware
is way too complex.

We use the middleware to map tokens to resource owners if available.
As the application currently expects a user in the string format we
use a subsequent middleware to map a resource owner to an id.