Make test pids and user ids valid in urls
The current format with spaces makes the PSR-15 bridge choke as a
request target cannot contain spaces.

It would normally be the responsibility of the web server to convert
spaces to %20 but we do not have such as thing during testing. Actual
pids should not contain spaces anyway.

Avoid the situation entirely by replacing spaces with -.