Refactor setup of Adgangsplatformen to use service providers
Reduce logic in app.php by instantiating services and middleware
provided by Adgangsplatformen in a dedicated provider which is always
used.

For testing the Adgangsplatformen OAuth provider is replaced with a 
test one which maps the provided access token to a resource owner 
with the same id. This is enabled by an update to the latest version 
of the package and replicates the current behavior.

This means we can drop some custom middleware. The test token checker
is now handled by the test OAuth provider. Resource owner to guid
mapping is now handled by a user resolver setup in the service
provider.