Use refactored version of Adgangsplatformen library
The refactored version contains some of the functionality that was
previously provided by the service. This means we can delete a bunch
of code.

Key changes are:

- All necessary providers and middleware are added to the application
by the provider. The bootstrap becomes much easier.
- The token validation middleware is registered using the auth
shorthand. Consequently we add all routes requiring authentication
to this. This also means you can check the version of the service
without authentication.
- The middleware now ensures that a domain specific user
object is returned for authenticated requests and not just a string.
Consequently we have to call the proper method to actually get the
user id in our controller.
- The middleware is also responsible for rejecting unauthenticated
requests so we can drop the TokenAccess middleware entirely.
- There is no need to define which driver Adgangsplatformen uses. For
testing it will automatically use the testing version.

Note that this currently depends on a branch of the Adgangsplatformen
library. This will be reverted to master once the changes there are
merged.