Skip contact info rule
It is not relevant for us to define contact in the specification.
This will depend on how and by who the service is deployed.

We could define this in a separate speccy.yaml file but that seems
execessive for a single option. We do not expect developers to run
Speccy locally anyway.