Move concrete specification versions to spec subfolder
We may end up supporting multiple versions which may be referenced
by external parties. To keep these in check we store them in the
spec subfolder and create a symlink to the latest version in the
root folder.

Update Dredd and Speccy tests to refer to the latest version.

In the future if we had matrix builds we could run previous versions 
of the spec against the current code try to determine backwards
incompatible changes.