Replace deprecated Google Cloud Action by GitHub with official action
actions/gcloud are deprecated and now no longer available. 
Consequently we have to move to the official actions released by
Google.

Setup is now handled by a single action which requires both service
account email and key, so we use that to setup the connection. *Then*
we can use the now available version of gcloud to configure Docker
and execute further actions.