Use stderr for Lumen error handling on Kubernetes
The current configuration uses default behavior which is to log to a 
file. This is not very useful for our GKE-based environments where we
want to send our logs to Stackdriver.

Stackdriver picks up data from stdout and stderr.

Consequently we switch to the stderr log channel which is provided by
default in Lumen.
https://github.com/laravel/lumen-framework/blob/6.x/config/logging.php#L74