Update migration controller to use UPDATE IGNORE for migrations
This way duplicate entries are ignored. In the current state they
result in duplicate key errors.

This query syntax. is not supporte by Illuminate/DB to we have to 
rely on writing the query by hand. There is also a different syntax 
for SQLite which we have to take into account for testing.