Improve handling of invalid pids
If we try to return a item id contained within a list stored in the
database we currently fail with a 422 Unprocessable Entity. This is not
very helpful as the client cannot do anything to address this.

To address this we change the process logic to mark a problem with the
list id by throwing a more general invalid argument exception. This can
then be caught and handled in different situations where this might 
occur:

- For situations where the item id is provided by the client we keep
responding with a 422 error.
- For situations where the item id is loaded from the database we log
the error and return any remaining items. 

The situation causing this issue is likely caused by a few faulty items
being migrated into the database in the initial stage of the project.
For v2 we have introduced stricter requirements for data formats. These
are being enforced for new data but cause problems with existing data.