Ensure that list items are always returned as an array
If an item has been removed then filter() will remove the null value
but preserve the keys of the remaining items. In JSON this will 
effectively turn the array into an object with keys.

Use values() to reset the keys.

Example: List with 3 items
Before: With entry 1 removed the remaining keys will be index 0 and 2.
After: With entry 1 remove the remaining keys will be 0 and 1.