Add support for health checks
Use the https://github.com/ans-group/laravel-health-check package to add
support for health checks. This is useful when adding external 
monitoring of the application.

The package used is somewhat popular, has explicit support for Lumen and
the somewhat old PHP version that we use.

The default configuration checks that we have acces to the database, a
working logging system and whatever environment variables that we need.
Tests show that it returns uncacheable response headers by default. 

That should be sufficient for an initial version of us.