Shift disk size from application to database
We are currently at less than 20% which trigger warnings. Bump the
disk size to ensure that we do not hit any limits.

The project currently has 5GB disk with 2GB used for the application
and 2GB used by the database. Bump the database by 1GB and decrease
the disk for the application by 1GB

Metrics show that we currently use <10MB for the application so
reducing it to 1GB should be OK.